requirements-dev.txt
-r requirements.txt
pytest==8.3
pytest-django==4.9
